requirements.txt
fastapi
uvicorn
torch
diffusers
sentence-transformers
pinecone
cloudinary
python-multipart
pydantic
python-dotenv
imageio
imageio-ffmpeg
opencv-python
accelerate